+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/dto/page/PageReqDto.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
-public class MainPageReqDto {
+public class PageReqDto {
 
     @AllArgsConstructor
     @Data
