--- tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/page/MainPageService.java
@@ -1,7 +1,0 @@
-package kr.ac.kumoh.illdang100.tovalley.service.page;
-import static kr.ac.kumoh.illdang100.tovalley.dto.page.MainPageRespDto.*;
-
-public interface MainPageService {
-
-    MainPageAllRespDto getMainPageAllData();
-}
