+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/page/PageService.java
@@ -2,9 +2,13 @@ package kr.ac.kumoh.illdang100.tovalley.service.page;
 
 import org.springframework.data.domain.Pageable;
 
-import static kr.ac.kumoh.illdang100.tovalley.dto.page.WaterPlaceDetailPageRespDto.*;
+import static kr.ac.kumoh.illdang100.tovalley.dto.page.PageRespDto.*;
 
-public interface WaterPlaceDetailPageService {
+public interface PageService {
+
+    MainPageAllRespDto getMainPageAllData();
 
     WaterPlaceDetailAllRespDto getWaterPlaceDetailPageAllData(Long waterPlaceId, Pageable pageable);
+
+    MyPageAllRespDto getMyPageAllData();
 }
