+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/trip_schedule/TripScheduleService.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.illdang100.tovalley.service.domain.trip_schedule;
+package kr.ac.kumoh.illdang100.tovalley.service.trip_schedule;
 
 import kr.ac.kumoh.illdang100.tovalley.domain.trip_schedule.TripSchedule;
